Web/OpenObjectivity/clicker2.0/requirements.txt
alembic==1.0.0
aniso8601==3.0.2
click==6.7
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-Migrate==2.2.1
flask-restplus==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
greenlet==0.4.15
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
monotonic==1.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
